Python/requirements.txt
google-generativeai
textract
mysql-connector-python
Flask
Werkzeug
python-dotenv